+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add support for normalized route names when using Lumen (#449)
+
 ## 2.11.1
 
 - Fix deprecation notice in route name extraction (#543)